+++ src/geodesy/wasmcontext.rs
@@ -1,11 +1,7 @@
-use crate::error::Result as GWResult;
-use geodesy_rs::authoring::{Context, OpConstructor, BUILTIN_ADAPTORS};
-use geodesy_rs::authoring::{Grid, Op, ParsedParameters};
-use geodesy_rs::prelude::*;
-use geodesy_rs::Error as RgError;
-use std::collections::BTreeMap;
-
 use super::operators::ACCESSORY_OPERATORS;
+use crate::error::Result as GWResult;
+use geodesy_rs::{authoring::*, Error as RgError};
+use std::{collections::BTreeMap, sync::Arc};
 
 // ----- T H E   W A S M  C T X   P R O V I D E R ---------------------------------
 // Modified from Rust Geodesy Minimal context to work with web inputs.
@@ -21,14 +17,14 @@ pub struct WasmContext {
     /// Instantiations of operators
     operators: BTreeMap<OpHandle, Op>,
     /// DataViews for external resources like grids
-    blobs: BTreeMap<String, Vec<u8>>,
+    grids: BTreeMap<String, Arc<dyn Grid>>,
 }
 
 const BAD_ID_MESSAGE: RgError = RgError::General("WasmContext: Unknown operator id");
 
 impl WasmContext {
-    pub fn set_blob(&mut self, key: &str, blob: Vec<u8>) -> GWResult<()> {
-        self.blobs.insert(key.to_string(), blob);
+    pub fn set_grid(&mut self, key: &str, grid: Arc<dyn Grid>) -> GWResult<()> {
+        self.grids.insert(key.to_string(), grid);
 
         Ok(())
     }
@@ -124,26 +120,22 @@ impl Context for WasmContext {
         ))
     }
 
-    fn get_blob(&self, name: &str) -> Result<Vec<u8>, RgError> {
-        // We can't use the file system in wasm so we must pre-parse data to `Vec<u8>` in `set_blob`
-        // ready to be consumed here.
-
-        // TODO: We should `remove` the resouce otherwise we're doubling the memory
-        // We'll need the `Context` trait to allow `get_blob(&mut self, name: &str) -> Result<Vec<u8>, RgError>`
-        if let Some(resource) = self.blobs.get(name) {
-            return Ok(resource.to_owned());
-        }
-
-        Err(RgError::NotFound(
-            name.to_string(),
-            ": User defined resource".to_string(),
+    fn get_blob(&self, _name: &str) -> Result<Vec<u8>, RgError> {
+        Err(RgError::General(
+            "Blob access by identifier not supported by the Wasm context provider",
         ))
     }
 
     /// Access grid resources by identifier
-    fn get_grid(&self, _name: &str) -> Result<Grid, RgError> {
-        Err(RgError::General(
-            "Grid access by identifier not supported by the Wasm context provider",
+    fn get_grid(&self, name: &str) -> Result<Arc<(dyn Grid + 'static)>, RgError> {
+        if let Some(grid) = self.grids.get(name) {
+            // It's a reference clone
+            return Ok(grid.clone());
+        }
+
+        Err(RgError::NotFound(
+            name.to_string(),
+            ": Grid resource".to_string(),
         ))
     }
 }