+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support `hgridshift` via `tidyProjString` helper which replaces `hgridshift` with `gridshift` [#23](https://github.com/Rennzie/geodesy-wasm/issues/23)
+- Update Rust Geodesy commit to use `ntv2` grid and multi-grid support
+  - Now supporting `@optional` and `@null` grids
 - `longlat` operator and aliases [#24](https://github.com/Rennzie/geodesy-wasm/issues/24)
 
 ## [0.3.0] - 2023-10-05